+++ FASTDOOM/r_sky.h
@@ -25,7 +25,7 @@
 // The sky map is 256*128*4 maps.
 #define ANGLETOSKYSHIFT 22
 
-extern int skytexture;
+extern short skytexture;
 
 // Called whenever the view size changes.
 void R_InitSkyMap(void);
